Множественные правки для первого релиза

+++ app/Http/Controllers/AuthorController.php
@@ -9,12 +9,11 @@ use App\Http\Requests\AuthorRequest;
 
 class AuthorController extends Controller
 {
-    // Вывод страницы автора
     public function showAuthorPage() {
         $author = Author::find(1);
 
         if ($author->avatar_path == null) {
-            $avatarUrl = asset('img/admin.jpeg');
+            $avatarUrl = asset('img/author.jpg');
         } else {
             $avatarUrl = asset('storage/'.$author->avatar_path);
         }
@@ -27,7 +26,7 @@ class AuthorController extends Controller
         $author = Author::find(1);
 
         if ($author->avatar_path == null) {
-            $avatarUrl = asset('img/admin.jpeg');
+            $avatarUrl = asset('img/author.jpg');
         } else {
             $avatarUrl = asset('storage/'.$author->avatar_path);
         }
@@ -35,28 +34,31 @@ class AuthorController extends Controller
         return view('admin.author', ['name' => $author['name'], 'info' => $author['info'], 'avatar' => $avatarUrl]);
     }
 
-    // Выведет форму обновления инфы об авторе
     public function showUpdateAuthorInfo() {
         $author = Author::find(1);
         return view('admin.author_update', ['name' => $author['name'], 'info' => $author['info']]);
     }
 
-    // Обновление инфы об авторе
     public function updateAuthorInfo(AuthorRequest $request) {
         $author = Author::find(1);
 
-        $oldAvatarPath = 'public/'.$author->avatar_path;
-        Storage::delete($oldAvatarPath);
-
         $author->name = $request['name'];
         $author->info = $request['info'];
-        $avatarPath = explode('/', $request->file('avatar')->store('public'));
-        $author->avatar_path = $avatarPath[1];
+
+        if ($request->avatar) {
+            if ($author->avatar_path != null) {
+                $oldAvatarPath = 'public/'.$author->avatar_path;
+                Storage::delete($oldAvatarPath);
+            }
+
+            $avatarPath = explode('/', $request->file('avatar')->store('public'));
+            $author->avatar_path = $avatarPath[1];
+        }
 
         $status = ($author->save()) ? 'Информация обновлена.' : 'Возникла ошибка.';
 
         $request->session()->flash('status', $status);
-        
+
         return redirect()->route('authorInfo');
     }
 }
